requirements.txt
streamlit
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
transformers
torch
accelerate
python-dotenv
pymupdf
pytesseract
pillow
python-docx
docx2txt